requirements.txt
mock
os
pytube
re
openai-whisper 
langchain 
langchain_openai
openai